feat: move assets related constants into fetch-assets
