+++ src/npc-cli/service/fetch-assets.js
@@ -1,5 +1,39 @@
-import { GEOMORPHS_JSON_FILENAME, assetsEndpoint, imgExt } from "src/const"; // 🚧 -> npc-cli/service/const
-import { isDevelopment } from "./generic";
+import { queryClient } from "./query-client";
+import { info, isDevelopment } from "./generic";
+
+/** See `npm run develop` */
+export const DEV_GATSBY_PORT = 8011;
+
+/** See `npm run ws-server` */
+export const DEV_EXPRESS_WEBSOCKET_PORT = 8012;
+
+export const DEV_ORIGIN = 'localhost';
+/**
+ * For local mobile debugging, e.g. via:
+ * `ifconfig | grep "inet " | grep -v 127.0.0.1`
+ */
+// export const DEV_ORIGIN = '192.168.16.66';
+
+/**
+ * - Parsed JSON stored at `static/assets/${ASSETS_META_JSON_FILENAME}`
+ * - Also a react-query `queryKey`.
+ */
+export const ASSETS_JSON_FILENAME = "assets.json";
+
+/**
+ * Gatsby serves `static/assets/*` in development as `/assets/*`.
+ * However, it can be slow to update.
+ * In development we serve assets directly to overcome this.
+ */
+export const assetsEndpoint = process.env.NODE_ENV === 'development'
+  ? `http://${DEV_ORIGIN}:${DEV_EXPRESS_WEBSOCKET_PORT}/dev-assets`
+  : '/assets'
+;
+
+export const GEOMORPHS_JSON_FILENAME = "geomorphs.json";
+
+export const imgExt = isDevelopment() ? 'png' : 'png.webp';
+
 
 /** @returns {Promise<Geomorph.GeomorphsJson>} */
 export async function fetchGeomorphsJson() {
@@ -20,3 +54,38 @@ export function getDecorSheetUrl() {
 function getDevCacheBustQueryParam() {
   return isDevelopment() ? `?v=${Date.now()}` : '';
 }
+
+export const WORLD_QUERY_FIRST_KEY = 'world';
+
+/**
+ * Dev-only event handling:
+ * > trigger component refresh on file change
+ */
+export function connectDevEventsWebsocket() {
+  const url = `ws://${DEV_ORIGIN}:${DEV_EXPRESS_WEBSOCKET_PORT}/dev-events`;
+  const wsClient = new WebSocket(url);
+  wsClient.onmessage = async (e) => {
+    info(`received websocket message:`, { url, data: e.data });
+
+    queryClient.refetchQueries({
+      predicate({ queryKey: [queryKey] }) {
+        return WORLD_QUERY_FIRST_KEY === queryKey;
+      },
+    });
+  };
+
+  wsClient.onopen = (e) => {
+    info(`${url} connected`);
+    wsAttempts = 0;
+  };
+  wsClient.onclose = (e) => {
+    info(`${url} closed: reconnecting...`);
+    if (++wsAttempts <= 5) {
+      setTimeout(connectDevEventsWebsocket, (2 ** wsAttempts) * 300);
+    } else {
+      info(`${url} closed: gave up reconnecting`);
+    }
+  };
+}
+
+let wsAttempts = 0;
